renamed FAQPanel to StandardFAQ, added default background colour to Accordion
