+++ client/src/components/StandardFAQ/StandardFAQ.yaml
@@ -1,4 +1,4 @@
-# FAQ keys for the FAQPanel story
+# FAQ keys for the StandardFAQ story
 
 question1_q:
   en: |
